+++ src/utils/http.rs
@@ -1,59 +1,73 @@
+use std::collections::HashMap;
 use std::error::Error;
 use std::fs::File;
 use std::io::Write;
 use std::path::Path;
-use std::str::FromStr;
 
-use futures::{Future, Stream};
-use hyper::{Client, Method, Request, Uri};
-use hyper::header::UserAgent;
-use hyper_rustls;
-use tokio_core::reactor;
+use cabot::{RequestBuilder, Client};
+use cabot::request::Request;
 
 pub fn download<'a, T: AsRef<str>>(url: &T, dest: &'a Path) -> Result<(), &'a Error> {
-    let mut core = reactor::Core::new().unwrap();
-    let uri = Uri::from_str(url.as_ref()).unwrap();
-
-    let client = Client::configure()
-        .connector(hyper_rustls::HttpsConnector::new(4, &core.handle()))
-        .build(&core.handle());
-
     let mut file = File::create(dest).expect("error creating file for download");
 
-    let work = client.get(uri).and_then(|res| {
-        res.body().for_each(
-            |chunk| file.write_all(&chunk).map_err(From::from),
-        )
-    });
+    let req = create_request(url);
+    let client = Client::new();
+    let res = client.execute(&req).unwrap();
 
-    core.run(work).unwrap();
-    return Ok(());
+    match res.status_code() {
+        301 | 302 => {
+            let headers = parse_headers(res.headers());
+            let location = headers.get("location").unwrap().as_str();
+            println!("download: redirect: {}", location);
+            return download(&location, dest);
+        }
+        _ => {}
+    };
+
+    file.write_all(res.body().unwrap()).unwrap();
+
+    Ok(())
 }
 
-pub fn fetch<'a, T: AsRef<str>>(url: &T) -> Result<Vec<u8>, &'a Error> {
-    let mut core = reactor::Core::new().unwrap();
-    let uri = Uri::from_str(url.as_ref()).unwrap();
+pub fn fetch<'a, T: AsRef<str>>(url: &T) -> Result<String, &'a Error> {
+    let req = create_request(url);
+    let client = Client::new();
+    let res = client.execute(&req).unwrap();
 
-    let client = Client::configure()
-        .connector(hyper_rustls::HttpsConnector::new(4, &core.handle()))
-        .build(&core.handle());
+    Ok(res.body_as_string().unwrap())
+}
 
-    let mut chunks = Vec::<u8>::new();
-    {
-        let mut req: Request = Request::new(Method::Get, uri);
-        let user_agent = format!("rust crate {} {}", env!("CARGO_PKG_NAME"), env!("CARGO_PKG_VERSION"));
-        req.headers_mut().set(UserAgent::new(user_agent));
-        let work = client.request(req).and_then(|res| {
-            res.body().for_each(
-                |chunk| chunks.write_all(&chunk).map_err(From::from),
-            )
-        });
-        core.run(work).unwrap();
+fn create_request<'a, T: AsRef<str>>(url: &T) -> Request {
+    RequestBuilder::new(url.as_ref())
+        .set_http_method("GET")
+        .add_header(&format!("User-Agent: {}", user_agent()))
+        .build()
+        .unwrap()
+}
+
+fn parse_headers (headers: Vec<&str>) -> HashMap<String, String> {
+    // HTTP RFC2616 says duplicate headers are fine
+    // but we deduplicate them here, which is fine for me for now
+    let mut map = HashMap::<String, String>::new();
+
+    for header in &headers {
+        let parts: Vec<&str> = header.splitn(2, ":").map(str::trim).collect();
+        if parts.len() == 2 {
+            let name = parts[0];
+            let normalised_name = str::to_lowercase(name);
+            map.insert(normalised_name, parts[1].to_string());
+        }
     }
 
-    let body = chunks.into_iter().collect();
+    return map;
+}
 
-    return Ok(body);
+fn user_agent () -> String {
+    format!(
+        "rust crate {} {}",
+        env!("CARGO_PKG_NAME"),
+        env!("CARGO_PKG_VERSION"),
+    )
 }
 
 #[cfg(test)]
@@ -63,11 +77,19 @@ mod tests {
     #[test]
     fn fetch_google() {
         match fetch(&"https://www.google.com") {
-            Ok(bytes) => {
-                let body = String::from_utf8(bytes).unwrap();
+            Ok(body) => {
                 assert!(body.contains("google"));
             }
             Err(_error) => assert!(false),
         }
+    }
+
+    #[test]
+    fn parse_sample_headers() {
+        let input = vec!["", "not a key-value pair", "Name: one", "name: two"];
+        let got = parse_headers(input);
+        let mut want = HashMap::<String, String>::new();
+        want.insert("name".to_string(), "two".to_string());
+        assert_eq!(got, want);
     }
 }