+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 * Windows: ensure ~/bin exists in PATH
 
+* pkg: [skaffold](https://github.com/GoogleCloudPlatform/skaffold): `sync` downloads into ~/bin
+
 ## [0.6.1] - 2018-03-14
 
 ### Fixed